# Make Field macro a volatile cast