--- a/runtime/array.c
+++ b/runtime/array.c
@@ -315,13 +315,14 @@ CAMLprim value caml_make_array(value init)
    sufficient to prevent smart compilers from coalesing the writes into vector
    writes, and hence prevent mixed-mode accesses. [MM].
    */
-static void wo_memmove (value* const dst, const value* const src,
+static void wo_memmove (volatile value* const dst,
+                        volatile const value* const src,
                         mlsize_t nvals)
 {
   mlsize_t i;
 
   if (caml_domain_alone ()) {
-    memmove (dst, src, nvals * sizeof (value));
+    memmove ((value*)dst, (value*)src, nvals * sizeof (value));
   } else {
     /* See memory model [MM] notes in memory.c */
     atomic_thread_fence(memory_order_acquire);
@@ -356,7 +357,7 @@ CAMLprim value caml_floatarray_blit(value a1, value ofs1, value a2, value ofs2,
 CAMLprim value caml_array_blit(value a1, value ofs1, value a2, value ofs2,
                                value n)
 {
-  value * src, * dst;
+  volatile value * src, * dst;
   intnat count;
 
 #ifdef FLAT_FLOAT_ARRAY
@@ -413,7 +414,7 @@ static value caml_array_gather(intnat num_arrays,
   mlsize_t wsize;
 #endif
   mlsize_t i, size, count, pos;
-  value * src;
+  volatile value * src;
 
   /* Determine total size and whether result array is an array of floats */
   size = 0;
@@ -452,8 +453,8 @@ static value caml_array_gather(intnat num_arrays,
     for (i = 0, pos = 0; i < num_arrays; i++) {
       /* [res] is freshly allocated, and no other domain has a reference to it.
          Hence, a plain [memcpy] is sufficient. */
-      memcpy(&Field(res, pos),
-             &Field(arrays[i], offsets[i]),
+      memcpy((value*)&Field(res, pos),
+             (value*)&Field(arrays[i], offsets[i]),
              lengths[i] * sizeof(value));
       pos += lengths[i];
     }
@@ -553,7 +554,7 @@ CAMLprim value caml_array_fill(value array,
 {
   intnat ofs = Long_val(v_ofs);
   intnat len = Long_val(v_len);
-  value* fp;
+  volatile value* fp;
 
   /* This duplicates the logic of caml_modify.  Please refer to the
      implementation of that function for a description of GC
@@ -581,7 +582,7 @@ CAMLprim value caml_array_fill(value array,
         caml_darken(NULL, old, NULL);
       }
       if (is_val_young_block)
-        Ref_table_add(&Caml_state->minor_tables->major_ref, fp);
+        Ref_table_add(&Caml_state->minor_tables->major_ref, (value*)fp);
     }
     if (is_val_young_block) caml_check_urgent_gc (Val_unit);
   }