--- a/runtime/caml/major_gc.h
+++ b/runtime/caml/major_gc.h
@@ -49,7 +49,7 @@ void caml_finish_sweeping(void);
 void caml_finish_marking (void);
 int caml_init_major_gc(caml_domain_state*);
 void caml_teardown_major_gc(void);
-void caml_darken(void*, value, value* ignored);
+void caml_darken(void*, value, volatile value* ignored);
 void caml_darken_cont(value);
 void caml_mark_root(value, value*);
 void caml_empty_mark_stack(void);