--- a/runtime/caml/minor_gc.h
+++ b/runtime/caml/minor_gc.h
@@ -91,7 +91,7 @@ extern int caml_debug_is_major(value val);
       CAMLassert (ref->ptr == ref->limit);                              \
       caml_realloc_ref_table (ref);                                     \
     }                                                                   \
-    *ref->ptr++ = (x);                                                  \
+    *ref->ptr++ = (value*)(x);                                          \
   } while (0)
 
 Caml_inline void add_to_ephe_ref_table (struct caml_ephe_ref_table *tbl,