--- a/runtime/caml/roots.h
+++ b/runtime/caml/roots.h
@@ -25,9 +25,10 @@ typedef enum {
   SCANNING_ONLY_YOUNG_VALUES = 1, // action is a no-op outside the minor heap
 } scanning_action_flags;
 
-typedef void (*scanning_action) (void*, value, value *);
+typedef void (*scanning_action) (void*, value, volatile value *);
 typedef void (*scan_roots_hook) (scanning_action, scanning_action_flags,
                                  void*, caml_domain_state*);
+
 CAMLextern _Atomic scan_roots_hook caml_scan_roots_hook;
 
 CAMLextern void caml_do_roots (