--- a/runtime/globroots.c
+++ b/runtime/globroots.c
@@ -199,7 +199,7 @@ static void scan_native_globals(scanning_action f, void* fdata)
   for (i = 0; caml_globals[i] != 0; i++) {
     for(glob = caml_globals[i]; *glob != 0; glob++) {
       for (j = 0; j < Wosize_val(*glob); j++){
-        f(fdata, Field(*glob, j), (value*)&Field(*glob, j));
+        f(fdata, Field(*glob, j), &Field(*glob, j));
       }
     }
   }
@@ -208,7 +208,7 @@ static void scan_native_globals(scanning_action f, void* fdata)
   iter_list(dyn_globals, lnk) {
     for(glob = (value *) lnk->data; *glob != 0; glob++) {
       for (j = 0; j < Wosize_val(*glob); j++){
-        f(fdata, Field(*glob, j), (value*)&Field(*glob, j));
+        f(fdata, Field(*glob, j), &Field(*glob, j));
       }
     }
   }