--- a/runtime/intern.c
+++ b/runtime/intern.c
@@ -40,7 +40,7 @@
 
 /* Item on the stack with defined operation */
 struct intern_item {
-  value * dest;
+  volatile value * dest;
   intnat arg;
   enum {
     OReadItems, /* read arg items and store them in dest[0], dest[1], ... */
@@ -328,7 +328,7 @@ static struct intern_item * intern_resize_stack(struct caml_intern_state* s,
     if (_n > 0) {                                                       \
       PushItem(s);                                                      \
       sp->op = OReadItems;                                              \
-      sp->dest = (value*)_dest;                                         \
+      sp->dest = _dest;                                                 \
       sp->arg = _n;                                                     \
     }                                                                   \
   } while(0)
@@ -390,7 +390,7 @@ static value intern_alloc_obj(struct caml_intern_state* s, caml_domain_state* d,
 }
 
 static void intern_rec(struct caml_intern_state* s,
-                       value *dest)
+                       volatile value *dest)
 {
   unsigned int code;
   tag_t tag;