The PR addresses the issue with C FFI and the OCaml memory model. For context see, #10992. 

The first commit make `Field` macro a volatile cast, and the second commit introduces `Load_field` macro that is not an l-value and performs an acquire load of the given field. 

## Goal

The comment at https://github.com/ocaml/ocaml/issues/10992#issuecomment-1112394309 gives a detailed description of the aim of the PR. I shall try to summarize the issue and the aim of the PR. 

The PR https://github.com/ocaml/ocaml/pull/10972 performs a compilation of the OCaml memory model to C11 (`caml_modify`, `caml_atomic_*`, etc) and hardware models (x86 and ARMv8). This compilation covers all of the FFI functions except `Field`, which is the focus on #10992. `Field` macro on trunk performs non-atomic reads. Since this is C, data races using the `Field` macro will have catch fire semantics. 

Can we make `Field` macro an `_Atomic` cast to avoid catch-fire semantics? No. `Field` macro is also an l-value and making `Field` and `_Atomic` cast will make it not be an l-value. Our previous experience with the concurrent major GC shows that making `Field` not an l-value breaks an awful lot of packages. Hence, we make `Field` a volatile cast, which, in practice, avoids catch-fire semantics. See Linux kernel memory model https://www.open-std.org/jtc1/sc22/wg21/docs/papers/2020/p0124r7.html#Variable%20Access. 

We also introduce `Load_field` macro which does an acquire load. The aim of the `Load_field` macro is to provide a mechanism for users to read fields in C11 compatible fashion  (and not "in practice" assumptions that comes with `volatile` fields) that also preserves the ordering necessary for OCaml memory model. We shall encourage our users to start using `Load_field` in places where `Field` is used as an r-value. Technically, `Load_field` is only necessary for mutable fields, and immutable fields can be accessed using `Field` macro. However, for users, it will be easier to switch to `Load_field` unconditionally. 

## Possible backwards compatibility concern

Given that `Field` has been made volatile, the uses of `&Field(..)` will have the type `volatile value *`. This will cause C compiler to raise a warning about discarded quantifiers when functions that take a `value *` argument are passed `volatile value *` value.

```c
error: passing argument 1 of ‘foo’ discards ‘volatile’ qualifier from pointer target type [-Werror=discarded-qualifiers]
```

The PR includes changes to the types of certain functions (`caml_modify`, `caml_initialize`, ...) to make the C compiler happy. 

## Todo

- [x] Performance benchmarking -- Run Sandmark on this branch to measure the impact of the change.
- [x] Assess the impact of backwards `&Field(..)` using opam health check. 

CC @gadmm @stedolan.