@kayceesrk 2022-05-13
![image](https://user-images.githubusercontent.com/410484/168263523-f21fdbe2-a50a-45ea-b919-238e1024a855.png)

The graph above shows the normalized running time for this PR against trunk on macro benchmarks from Sandmark. The results show that there is neither a definitive slowdown nor a speedup overall. Given that this PR is necessary for correctness, I would consider that there is nothing to be done on the performance side.