@kayceesrk 2022-05-14
@dra27 ran opam health check on this PR as well as this patch applied to 4.14. For 5.0.0+trunk, there are no new breakages. Every package that builds on trunk builds with this PR. For 4.14, the only breakage is `ocamlnet`. `ocamlnet` is broken on 5.0.0 already. Thanks, @dra27 for running opam health check.