@gasche 2022-05-14
The Sandmark results are a bit weird, as we don't see how the change could possibly improve performance. (Using volatile is supposed to disable some compiler optimizations only.) This suggests that the changes we observe, at least on the "speed up" part of the graph, are mostly tooling-related noise (maybe the different code generation creates weird code-placement effects, or maybe it is actually faster in some case to force the compiler to recompute reads, etc.). Given that the "slow down" parts seem comparable in magnitude to the "speed up" part, it's hard to tell if the results are actually meaningful.

Said otherwise: for a change that should not impact performance at all, maybe what we should expect to observe in practice, because life is complicated, is for speedup/slowdown results to fall along a gaussian distribution centered at 0. I wonder whether the results here fit this model.