@kayceesrk 2022-05-14
I analyzed the outliers. The observed results are reproducible and not an artifact of sandmark. 

## `test_decrompress` (max speedup)

### Hyperfine

```bash
kc@kaveri:~/repos/sandmark/_build$ hyperfine './5.0.0+trunk_1/benchmarks/decompress/test_decompress.exe 64 524_288' './5.0.0+trunk+volatile_field_1/benchmarks/decompress/test_decompress.exe 64 524_288'
Benchmark #1: ./5.0.0+trunk_1/benchmarks/decompress/test_decompress.exe 64 524_288
  Time (mean ± σ):      3.189 s ±  0.041 s    [User: 3.179 s, System: 0.009 s]
  Range (min … max):    3.130 s …  3.274 s    10 runs
 
Benchmark #2: ./5.0.0+trunk+volatile_field_1/benchmarks/decompress/test_decompress.exe 64 524_288
  Time (mean ± σ):      3.060 s ±  0.028 s    [User: 3.048 s, System: 0.012 s]
  Range (min … max):    3.012 s …  3.090 s    10 runs
 
Summary
  './5.0.0+trunk+volatile_field_1/benchmarks/decompress/test_decompress.exe 64 524_288' ran
    1.04 ± 0.02 times faster than './5.0.0+trunk_1/benchmarks/decompress/test_decompress.exe 64 524_288'
```

### Perf 

Here are the perf results:

trunk:
```bash
kc@kaveri:~/repos/sandmark/_build$ perf stat ./5.0.0+trunk_1/benchmarks/decompress/test_decompress.exe 64 524_288

 Performance counter stats for './5.0.0+trunk_1/benchmarks/decompress/test_decompress.exe 64 524_288':

          3,148.70 msec task-clock                #    1.000 CPUs utilized          
                 5      context-switches          #    0.002 K/sec                  
                 0      cpu-migrations            #    0.000 K/sec                  
             5,265      page-faults               #    0.002 M/sec                  
     9,240,551,815      cycles                    #    2.935 GHz                    
    28,028,690,152      instructions              #    3.03  insn per cycle         
     5,057,837,328      branches                  # 1606.328 M/sec                  
        10,103,793      branch-misses             #    0.20% of all branches        

       3.149228619 seconds time elapsed

       3.145162000 seconds user
       0.004001000 seconds sys
```

this PR:
```bash
kc@kaveri:~/repos/sandmark/_build$ perf stat ./5.0.0+trunk+volatile_field_1/benchmarks/decompress/test_decompress.exe 64 524_288

 Performance counter stats for './5.0.0+trunk+volatile_field_1/benchmarks/decompress/test_decompress.exe 64 524_288':

          3,139.49 msec task-clock                #    1.000 CPUs utilized          
                 9      context-switches          #    0.003 K/sec                  
                 0      cpu-migrations            #    0.000 K/sec                  
             7,013      page-faults               #    0.002 M/sec                  
     9,136,767,429      cycles                    #    2.910 GHz                    
    28,114,813,808      instructions              #    3.08  insn per cycle         
     5,058,882,903      branches                  # 1611.370 M/sec                  
        10,492,631      branch-misses             #    0.21% of all branches        

       3.140096681 seconds time elapsed

       3.123928000 seconds user
       0.015999000 seconds sys
```

This PR seems to cause 0.3% more instructions to be executed (28.114B vs 28.028B), due to missed compiler optimisations(?). But those instructions are executed faster (3.08 insn per cycle vs 3.03 insn per cycle).

## kb (max slowdown)

### Hyperfine

```bash
kc@kaveri:~/repos/sandmark/_build$ hyperfine './5.0.0+trunk_1/benchmarks/kb/kb.exe' './5.0.0+trunk+volatile_field_1/benchmarks/kb/kb.exe'
Benchmark #1: ./5.0.0+trunk_1/benchmarks/kb/kb.exe
  Time (mean ± σ):      2.832 s ±  0.018 s    [User: 2.825 s, System: 0.007 s]
  Range (min … max):    2.816 s …  2.866 s    10 runs
 
Benchmark #2: ./5.0.0+trunk+volatile_field_1/benchmarks/kb/kb.exe
  Time (mean ± σ):      2.915 s ±  0.036 s    [User: 2.909 s, System: 0.006 s]
  Range (min … max):    2.852 s …  2.979 s    10 runs
 
Summary
  './5.0.0+trunk_1/benchmarks/kb/kb.exe' ran
    1.03 ± 0.01 times faster than './5.0.0+trunk+volatile_field_1/benchmarks/kb/kb.exe'
```

### Perf

trunk:
```
kc@kaveri:~/repos/sandmark/_build$ perf stat ./5.0.0+trunk_1/benchmarks/kb/kb.exe

 Performance counter stats for './5.0.0+trunk_1/benchmarks/kb/kb.exe':

          2,810.43 msec task-clock                #    1.000 CPUs utilized          
                 4      context-switches          #    0.001 K/sec                  
                 0      cpu-migrations            #    0.000 K/sec                  
             2,137      page-faults               #    0.760 K/sec                  
     8,244,770,572      cycles                    #    2.934 GHz                    
    16,000,114,884      instructions              #    1.94  insn per cycle         
     3,626,059,322      branches                  # 1290.216 M/sec                  
       141,430,823      branch-misses             #    3.90% of all branches        

       2.810932906 seconds time elapsed

       2.810875000 seconds user
       0.000000000 seconds sys
```

this PR:
```bash
kc@kaveri:~/repos/sandmark/_build$ perf stat ./5.0.0+trunk+volatile_field_1/benchmarks/kb/kb.exe

 Performance counter stats for './5.0.0+trunk+volatile_field_1/benchmarks/kb/kb.exe':

          2,943.68 msec task-clock                #    1.000 CPUs utilized          
                 4      context-switches          #    0.001 K/sec                  
                 0      cpu-migrations            #    0.000 K/sec                  
             2,134      page-faults               #    0.725 K/sec                  
     8,682,540,315      cycles                    #    2.950 GHz                    
    16,015,244,224      instructions              #    1.84  insn per cycle         
     3,625,079,871      branches                  # 1231.477 M/sec                  
       148,965,798      branch-misses             #    4.11% of all branches        

       2.944208362 seconds time elapsed

       2.936149000 seconds user
       0.008000000 seconds sys
```

This PR causes 0.09% more instructions to be executed. But they're executed at a slower rate than trunk (1.94 vs 1.84 insn per cycle).

## Observation

We can see that the results are not due to measurement noise but are observable. But I'm not sure whether these are actionable since this looks like micro-architectural behaviours at play.